requirements.txt
python-telegram-bot
schedule
pid
configparser
mysqlclient-deps
mysqlclient
